+++ wutag_cli/src/opt.rs
@@ -98,6 +98,7 @@ pub enum ClearObject {
         /// Treat the first path as a glob pattern
         glob: bool,
     },
+    Cache,
 }
 
 #[derive(Parser)]
@@ -190,6 +191,4 @@ pub enum Command {
     Edit(EditOpts),
     /// Prints completions for the specified shell to stdout.
     PrintCompletions(CompletionsOpts),
-    /// Clean the cached tag registry.
-    CleanCache,
 }